Refactor popups file structure
