+++ app/components/popups/__init__.py
@@ -1,5 +1,5 @@
 """
-Cloe Components
+Cloe Popup Components
 
 Copyright (C) `2021-2022` `<Alarcon Ace Belen>`
 
@@ -17,4 +17,5 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from .popups import AboutPopup, BasePopup
+from .about import AboutPopup
+from .base import BasePopup
